+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7] - 2026-05-11
+
+- Rar wrapper (reader) in C++ around unarr
+- Rar writer exporting Rar class of RarLib
+
 ## [0.0.6] - 2026-05-11
 
 ### Added
@@ -60,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic initialisation of repository
 - Basic function, converting each file of a PDF into JPG, to test GitHub CI configuration - this function will be removed
 
-[unreleased]: https://github.com/DostLeFan/Booker/compare/v0.0.6...HEAD
+[unreleased]: https://github.com/DostLeFan/Booker/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/DostLeFan/Booker/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/DostLeFan/Booker/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/DostLeFan/Booker/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/DostLeFan/Booker/compare/v0.0.3...v0.0.4